+++ UIComponents.Roslyn/UIComponents.Roslyn.Generation.Tests/Snapshots/DependencyAugmentGeneratorSnapshotTests.It_Adds_Override_Keyword_When_Used_On_UIComponent.01UIComponentWithNoOwnDependencies.Dependencies.g.verified.cs
@@ -12,13 +12,13 @@ using UnityEngine.UIElements;
 
 public partial class UIComponentWithNoOwnDependencies
 {
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     private static readonly IDependency[] UIC_Dependencies = new IDependency[] {
         UIComponents.DependencyInjection.Dependency.SingletonFor<UIComponents.IAssetResolver, UIComponents.ResourcesAssetResolver>(),
         UIComponents.DependencyInjection.Dependency.SingletonFor<UIComponents.ILogger, UIComponents.DebugLogger>()
     };
 
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     public override IEnumerable<IDependency> GetDependencies()
     {
         return UIC_Dependencies;