+++ UIComponents.Roslyn/UIComponents.Roslyn.Generation.Tests/Snapshots/LayoutAugmentGeneratorSnapshotTests.It_Takes_AssetPrefixAttribute_Into_Account.verified.cs
@@ -11,7 +11,7 @@ using System.Threading.Tasks;
 
 public partial class LayoutTestComponent
 {
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     protected override Task<VisualTreeAsset> UIC_StartLayoutLoad()
     {
         return AssetResolver.LoadAsset<VisualTreeAsset>("UI/Components/LayoutTestComponent");