+++ UIComponents.Roslyn/UIComponents.Roslyn.Generation.Tests/Snapshots/QueryAugmentGeneratorSnapshotTests.It_Handles_Common_Namespaces.verified.cs
@@ -14,7 +14,7 @@ namespace MyLibrary.GUI
 {
 public partial class MyComponent
 {
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     protected override void UIC_PopulateQueryFields()
     {
         // element