+++ UIComponents.Roslyn/UIComponents.Roslyn.Generation.Tests/Snapshots/QueryAugmentGeneratorSnapshotTests.It_Handles_Lists.verified.cs
@@ -11,7 +11,7 @@ using System.CodeDom.Compiler;
 
 public partial class ListQueryComponent
 {
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     protected override void UIC_PopulateQueryFields()
     {
         // elements