+++ UIComponents.Roslyn/UIComponents.Roslyn.Generation.Tests/Snapshots/QueryAugmentGeneratorSnapshotTests.It_Handles_Namespace_And_Nested_Type.verified.cs
@@ -15,7 +15,7 @@ public partial class ParentClass
 {
 public partial class NestedComponent
 {
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     protected override void UIC_PopulateQueryFields()
     {
         // field