+++ UIComponents.Roslyn/UIComponents.Roslyn.Generation.Tests/Snapshots/RegistersEventCallbackGeneratorSnapshotTests.It_Handles_Interface_Inheritance.verified.cs
@@ -10,14 +10,14 @@ using UnityEngine.UIElements;
 
 public partial class EventHandler
 {
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     protected override void UIC_RegisterEventCallbacks()
     {
         RegisterCallback<MyEvent>(OnMyEvent);
         RegisterCallback<OtherEvent>(OnOtherEvent);
     }
 
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     protected override void UIC_UnregisterEventCallbacks()
     {
         UnregisterCallback<MyEvent>(OnMyEvent);