+++ UIComponents.Roslyn/UIComponents.Roslyn.Generation.Tests/Snapshots/UxmlAugmentGeneratorSnapshotTests.Allows_Setting_Default_Value.verified.cs
@@ -10,17 +10,17 @@ using UnityEngine.UIElements;
 
 public partial class ComponentWithDefaultValueTraits
 {
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     public new partial class UxmlFactory : UxmlFactory<ComponentWithDefaultValueTraits, UxmlTraits> {}
 
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     public new partial class UxmlTraits : VisualElement.UxmlTraits
     {
         UxmlStringAttributeDescription m_Description = new UxmlStringAttributeDescription { name = "description" };
         UxmlIntAttributeDescription m_Lives = new UxmlIntAttributeDescription { name = "lives" };
         UxmlEnumAttributeDescription<Some.Place.Where.Enum.Is.TheEnum> m_MyValue = new UxmlEnumAttributeDescription<Some.Place.Where.Enum.Is.TheEnum> { name = "custom-value" };
 
-        [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+        [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
         public override void Init(VisualElement ve, IUxmlAttributes bag, CreationContext cc)
         {
             base.Init(ve, bag, cc);