+++ UIComponents.Roslyn/UIComponents.Roslyn.Generation.Tests/Snapshots/UxmlAugmentGeneratorSnapshotTests.Generates_Both_Traits_And_UxmlFactory.verified.cs
@@ -10,28 +10,28 @@ using UnityEngine.UIElements;
 
 public partial class ComponentWithUxmlNameAndTraits
 {
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     public new partial class UxmlFactory : UxmlFactory<ComponentWithUxmlNameAndTraits, UxmlTraits>
     {
-        [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+        [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
         public override string uxmlName
         {
             get { return "AwesomeUxmlName"; }
         }
 
-        [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+        [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
         public override string uxmlQualifiedName
         {
             get { return uxmlNamespace + "." + uxmlName; }
         }
     }
 
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     public new partial class UxmlTraits : VisualElement.UxmlTraits
     {
         UxmlBoolAttributeDescription m_Value = new UxmlBoolAttributeDescription { name = "value" };
 
-        [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+        [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
         public override void Init(VisualElement ve, IUxmlAttributes bag, CreationContext cc)
         {
             base.Init(ve, bag, cc);