+++ UIComponents.Roslyn/UIComponents.Roslyn.Generation.Tests/Snapshots/UxmlAugmentGeneratorSnapshotTests.Generates_String_Traits.verified.cs
@@ -10,16 +10,16 @@ using UnityEngine.UIElements;
 
 public partial class StringComponentWithUsing
 {
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     public new partial class UxmlFactory : UxmlFactory<StringComponentWithUsing, UxmlTraits> {}
 
-    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+    [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
     public new partial class UxmlTraits : VisualElement.UxmlTraits
     {
         UxmlStringAttributeDescription m_FieldTrait = new UxmlStringAttributeDescription { name = "field-trait" };
         UxmlStringAttributeDescription m_PropertyTrait = new UxmlStringAttributeDescription { name = "property-trait" };
 
-        [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-alpha.6")]
+        [GeneratedCode("UIComponents.Roslyn.Generation", "1.0.0-beta.0")]
         public override void Init(VisualElement ve, IUxmlAttributes bag, CreationContext cc)
         {
             base.Init(ve, bag, cc);