+++ UIComponents.Roslyn/UIComponents.Roslyn.Generation/Constants.cs
@@ -3,7 +3,7 @@
     internal static class Constants
     {
         public const string ToolName = "UIComponents.Roslyn.Generation";
-        public const string Version = "1.0.0-alpha.6";
+        public const string Version = "1.0.0-beta.0";
         public const string GeneratedCodeAttribute = "[GeneratedCode(\"" + ToolName + "\", \"" + Version + "\")]";
     }
 }